Weight GGX diffuse-specular sampling
